Add functional tests for the + operator.
Add functional tests for the + operator.

Extend implementation of name/attribute transformers so that they
actually transform variables into the valid rtseq equivalents instead of
forcing everyone else to go through the resources dictionary.

Add a small helper to join all the elements from an attribute so that we
can get the full a.b.value in one call.