Add Nivs_Param decorator class to describe non-default parameters
The Nivs_Param class __init__ during function definition, the the __call__ is invoked at call time. In the function_transformer we do two passes on parameters, first on the signature of the function to get the names, then on the decorators to get their descriptions and update the dictionary.

When running in python mode, the parameter descriptions will instead be saved as a parameter of the wrapped function, which are then used in _reconstruct_parameters() to make copies when passing by value.