Add timing libraries sequences.
Adds wait, wait_until_ms_multiple, wait_until_us_multiple, and wait_until_settled.
Implements the python version of functions needed (seqtimeXX, quotiend)
Small refactor in parameters so that if a parameter is passed by ref but it's just a base type (i.e. int, float) we raise an exception.

Testing is pretty basic because every test we add introduces a lot of time to our test runs because we actually have to wait in real time.