Merged PR 98: Support having pydoc in functiondefs.
Pydocs are special strings instead of comments, so they do show up in the ast. So create a validation helper to check if something looks like a docstring and ignore it from validation in try/finally blocks and functiondefs. If it's found in the middle of code (i.e. it's an expression) then we do nothing. In the future we could choose to insert a comment block if we wanted to.