Merged PR 106: Hardening for if and while transformers.
Adds validations to if and while transformers.
A subroutine_call test broke because of this (it was returning inside an if) so since I already had to touch the file, I went and did some cleaning on it. Also added a few more validations. I don't count this as full hardening of call transformer though, that will come later.